+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ## (IN PROGRESS)
 
+* [UIBULKED-3](https://issues.folio.org/browse/UIBULKED-3) Add Identifier and Query options to the landing page
 * [UIBULKED-1](https://issues.folio.org/browse/UIBULKED-1) Landing page
 * [UIBULKED-2](https://issues.folio.org/browse/UIBULKED-2) Project Setup